Generate new scanner model definitions
